Create ActivityListView file and create the List using Activity Cell
