+++ solutions/devsprint-julio-fernandes-6/FinanceApp/Screens/Home/ActivityCellView.swift
@@ -62,7 +62,6 @@ class ActivityCellView: UITableViewCell {
     
     override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
         super.init(style: style, reuseIdentifier: reuseIdentifier)
-        
         setupView()
     }
